+++ lib/src/core/handler_proxy.dart
@@ -6,24 +6,39 @@ part of core;
 /// Example:
 ///
 /// ```dart
-/// proxy = Proxy('~/foobar/*', 'http://example.com');
+/// proxy = Proxy('GET', '~/foobar/*', 'http://example.com');
 /// proxy.register(pipeline);
 /// ```
 ///
-/// Requests for '~/foobar/abc/def' will return the response from
-/// "http://example.com/foobar/abc/def".
+/// GET requests for '~/foobar/abc/def' will return the response from
+/// sending a request to "http://example.com/abc/def".
 
 class Proxy {
   //================================================================
   // Constructors
 
   //----------------------------------------------------------------
-  /// Constructor
+  /// Create a request handler that proxies requests to another URL.
+  ///
+  /// This request handler is used to handle requests that match the [method]
+  /// and [pattern]. The pattern must end with a "*" (e.g. "~/*" or
+  /// "~/foobar/*"). The path that matches the "*" is the sub-path.
+  ///
+  /// When handling a request, it will make a proxy request to a URL that is
+  /// made up of the [proxy] with the sub-path appended. The response from that
+  /// proxy request is forwarded back as the response.
+  ///
+  /// A "Via" header is always added to the proxy request. The value of [via]
+  /// is used, unless it is null or the empty string. Otherwise, a default
+  /// value is used. A "Via" header containing the same value is added to the
+  /// response, unless [includeViaHeaderInResponse] is false.
 
   Proxy(this.method, String pattern, String proxy,
-      {List<String> ignoreNotFound,
-      this.requestBlockHeaders,
-      this.responseBlockHeaders}) {
+      {String via,
+      this.includeViaHeaderInResponse = true,
+      @deprecated Iterable<String> requestBlockHeaders,
+      @deprecated Iterable<String> responseBlockHeaders})
+      : _via = (via != null && via.isNotEmpty) ? via : _viaDefault {
     if (method != 'GET' && method != 'HEAD') {
       throw ArgumentError.value(
           method, 'method', 'only GET and HEAD supported');
@@ -37,12 +52,21 @@ class Proxy {
       throw ArgumentError.value(pattern, 'pattern', 'does not end with "*"');
     }
 
-    _pathPrefix = pattern.substring(2, pattern.length - 2);
+    _pathPrefix =
+        pattern == '~/*' ? '' : pattern.substring(2, pattern.length - 2);
     _proxyHost = proxy;
 
-    if (ignoreNotFound != null) {
-      for (var path in ignoreNotFound) {
-        _ignoreNotFound.add('$proxy/$path');
+    // Store lower case versions of headers to block.
+
+    if (requestBlockHeaders != null) {
+      for (final name in requestBlockHeaders) {
+        this.requestBlockHeaders.add(name.toLowerCase());
+      }
+    }
+
+    if (responseBlockHeaders != null) {
+      for (final name in responseBlockHeaders) {
+        this.responseBlockHeaders.add(name.toLowerCase());
       }
     }
   }
@@ -51,17 +75,36 @@ class Proxy {
   // Static constants
 
   /// Headers in the request which are never passed through to the target.
+  ///
+  /// Important: these must be in all lowercase, otherwise matching won't work.
 
-  static const List<String> requestHeadersNeverPass = ['host', 'connection'];
+  static const List<String> requestHeadersNeverPass = [
+    'host', // host will be recreated for the request
+    'connection', // HTTP/1.1 keep-alive is not supported
+    'keep-alive',
+    'te',
+    'upgrade',
+    'upgrade-insecure-requests', // should this be included or not?
+    'proxy-authorization',
+  ];
 
   /// Headers in the response which are never passed through to the client.
+  ///
+  /// Important: these must be in all lowercase, otherwise matching won't work.
 
   static const List<String> responseHeadersNeverPass = [
-    'content-type',
-    'x-content-type-options',
-    'x-frame-options',
-    'x-xss-protection'
+    'content-type', // content-type will be recreated for the response
+    'connection', // HTTP/1.1 keep-alive is not supported
+    'keep-alive',
+    'transfer-encoding',
+    'trailer',
+    'proxy-authenticate',
   ];
+
+  // Default value to use for the "Via" header when forwarding the request.
+  // Used when no value of via was provided.
+
+  static const String _viaDefault = '1.1 woomera_proxy';
 
   //================================================================
   // Members
@@ -74,11 +117,13 @@ class Proxy {
 
   String _pathPrefix;
 
-  /// Paths that will be ignored if the target returns HTTP Status 404.
-  ///
-  /// No warnings are logged for these.
+  /// Value for the "Via" header.
 
-  final List<String> _ignoreNotFound = [];
+  final String _via;
+
+  /// Indicates if a Via header is added to the response.
+
+  final bool includeViaHeaderInResponse;
 
   /// Additional request headers which are not passed through to the target.
   ///
@@ -87,7 +132,7 @@ class Proxy {
   ///
   /// Set this value in the constructor.
 
-  final List<String> requestBlockHeaders;
+  final List<String> requestBlockHeaders = [];
 
   /// Additional response headers which are not passed through to the client.
   ///
@@ -96,7 +141,7 @@ class Proxy {
   ///
   /// Set this value in the constructor.
 
-  final List<String> responseBlockHeaders;
+  final List<String> responseBlockHeaders = [];
 
   //================================================================
   // Methods
@@ -156,71 +201,199 @@ class Proxy {
   ///
 
   Future<Response> handleGetOrHead(Request req) async {
-    // Determine the target URI
+    _logProxy.fine('[${req.id}] $method ${req.requestPath()}');
+
+    // Determine the target URI for the proxy request
 
     final targetUrl = _targetUri(req);
-    _logProxy.fine('[${req.id}] $method $targetUrl');
+    _logProxyRequest.finer('[${req.id}] $method $targetUrl');
 
     try {
-      // Determine the request headers to send to the target
-      //
-      // Note: the 'http' package does not support multiple headers with the
-      // same name (which are possible in HTTP).
+      // Determine headers for the proxy request
 
-      final passHeaders = <String, String>{};
+      final _prHeaders = await _proxyRequestHeaders(req, targetUrl);
 
-      final u = Uri.parse(targetUrl);
-      passHeaders['host'] = (u.hasPort) ? '${u.host}:${u.port}' : u.host;
+      // Perform the proxy request
 
-      req.headers.forEach((headerName, values) {
-        if (!(requestHeadersNeverPass.contains(headerName) ||
-            (requestBlockHeaders?.contains(headerName) ?? false))) {
-          if (values.length == 1) {
-            passHeaders[headerName] = values.first;
-          } else {
-            _logProxy.warning('request header not a single value: $headerName');
-          }
-        }
-      });
-
-      // Perform request for the target
-
-      final targetResponse = await http.get(targetUrl, headers: passHeaders);
+      final targetResponse = await http.get(targetUrl, headers: _prHeaders);
       assert(targetResponse != null);
 
-      if (targetResponse.statusCode != HttpStatus.ok &&
-          targetResponse.statusCode != HttpStatus.notModified) {
-        if (!(targetResponse.statusCode == HttpStatus.notFound &&
-            _ignoreNotFound.contains(targetUrl))) {
-          _logProxy.warning('$targetUrl: status ${targetResponse.statusCode}');
-        }
-      }
+      // Produce the response from the response of the proxy request
 
-      // Pass the target's response back as the response
-
-      final contentType = (targetResponse.headers.containsKey('content-type'))
-          ? ContentType.parse(targetResponse.headers['content-type'])
-          : ContentType.binary;
-
-      // Use the response from the target as the response
-
-      final resp = ResponseBuffered(contentType)
-        ..status = targetResponse.statusCode;
-
-      for (var headerName in targetResponse.headers.keys) {
-        if (!(responseHeadersNeverPass.contains(headerName) ||
-            (responseBlockHeaders?.contains(headerName) ?? false))) {
-          // Not one of the special headers: copy it to the response
-          resp.headerAdd(headerName, targetResponse.headers[headerName]);
-        }
-      }
-
-      resp.write(targetResponse.body);
-      return resp;
+      return await _produceResponse(req, targetResponse);
     } catch (e) {
       final proxyException = ProxyHandlerException(targetUrl, e);
       _logProxy.fine('[${req.id}] $proxyException');
       throw proxyException;
     }
+  }
+
+  //----------------
+
+  Future<Map<String, String>> _proxyRequestHeaders(
+      Request req, String targetUrl) async {
+    // Determine the request headers to send to the target
+    //
+    // Note: the 'http' package does not support multiple headers with the
+    // same name (which are possible in HTTP).
+
+    final passHeaders = <String, String>{};
+
+    // New "Host" header for outgoing request
+
+    final u = Uri.parse(targetUrl);
+    final _newHost = (u.hasPort) ? '${u.host}:${u.port}' : u.host;
+    passHeaders['host'] = _newHost;
+    _logProxyRequest.finest('[${req.id}] + host=$_newHost');
+
+    // Identify any connection headers that must not be passed through
+
+    // ignore: prefer_collection_literals
+    final requestConnectionExclude = Set<String>();
+
+    final c = req.headers['connection'];
+    if (c != null) {
+      for (final v in c) {
+        for (final name in v.split(',')) {
+          requestConnectionExclude.add(name.trim().toLowerCase());
+        }
+      }
+    }
+
+    // Headers from incoming request to outgoing request
+
+    req.headers.forEach((key, values) {
+      final headerName = key.toLowerCase();
+
+      if ((requestHeadersNeverPass.contains(headerName) ||
+              requestConnectionExclude.contains(headerName.toLowerCase()) ||
+              (requestBlockHeaders.contains(headerName))) ||
+          headerName == 'via') {
+        // Do not pass through header
+        _logProxyRequest.finest('[${req.id}] - $headerName=${values.first}');
+      } else {
+        // Pass through header
+        if (values.length == 1) {
+          final value = values.first;
+          _logProxyRequest.finest('[${req.id}]   $headerName=$value');
+          passHeaders[headerName] = value;
+        } else {
+          _logProxyRequest
+              .warning('[${req.id}] header not a single value: $headerName');
+        }
+      }
+    });
+
+    /// A "Via" header is _always_ added to forwarded requests.
+    /// It is mandatory according to
+    /// [RFC 7230](https://tools.ietf.org/html/rfc7230#section-5.7.1).
+
+    final _newVia = req.headers['via'] != null
+        ? '${req.headers['via'].where((s) => s.isNotEmpty).join(', ')}, $_via'
+        : _via;
+
+    passHeaders['via'] = _newVia;
+    _logProxyRequest.finest('[${req.id}] + via=$_newVia');
+
+    return passHeaders;
+  }
+
+  //----------------
+
+  Future<Response> _produceResponse(
+      Request req, http.Response targetResponse) async {
+    // Log response status
+
+    _logProxyResponse.finer('[${req.id}] status=${targetResponse.statusCode}');
+
+    // Pass the target's response back as the response
+
+    // Extract the Content-Type header to be used to create ResponseStream.
+    //
+    // If there is no Content-Type header, assume application/binary, since
+    // ResponseStream must have a content type.
+
+    final contentType = (targetResponse.headers.containsKey('content-type'))
+        ? ContentType.parse(targetResponse.headers['content-type'])
+        : ContentType.binary;
+
+    _logProxyResponse.finest('[${req.id}] + content-type=$contentType');
+
+    // Identify any connection headers that must not be passed through
+
+    // ignore: prefer_collection_literals
+    final responseConnectionExclude = Set<String>();
+
+    final c2 = targetResponse.headers['connection'];
+    if (c2 != null) {
+      for (final name in c2.split(',')) {
+        responseConnectionExclude.add(name.trim().toLowerCase());
+      }
+    }
+
+    // Use the response from the target as the response
+    //
+    // Must use a _ResponseStream_ because the response body needs to be treated
+    // as binary data and not a string in any particular encoding. Otherwise,
+    // any encoding changes will not match the content-length header.
+
+    final resp = ResponseStream(contentType)
+      ..status = targetResponse.statusCode
+      ..headerRemoveAll();
+
+    // Above removes all default headers, since some/all of them might be
+    // present in the proxy response. Only the ones from the proxy response will
+    // appear in the response.
+    //
+    // Cannot blindly use resp.headerAdd since that could produce multiple
+    // headers as well as we need to avoid this bug:
+    //   https://github.com/dart-lang/sdk/issues/43627
+
+    // Forward headers from proxy response to the response
+
+    int _size;
+    for (var key in targetResponse.headers.keys) {
+      final headerName = key.toLowerCase();
+      final headerValue = targetResponse.headers[key];
+
+      if (responseHeadersNeverPass.contains(headerName) ||
+          responseConnectionExclude.contains(headerName.toLowerCase()) ||
+          (responseBlockHeaders.contains(headerName)) ||
+          headerName == 'via') {
+        // Do not pass back header
+        _logProxyResponse.finest('[${req.id}] - $headerName=$headerValue');
+      } else {
+        // Pass header back to client
+        if (headerName == 'content-length') {
+          _size = int.parse(headerValue);
+        }
+        resp.headerAdd(headerName, headerValue);
+        _logProxyResponse.finest('[${req.id}]   $headerName=$headerValue');
+      }
+    }
+
+    // Via header
+
+    if (includeViaHeaderInResponse) {
+      /// A "Via" header is optional in forwarded responses according to
+      /// [RFC 7230](https://tools.ietf.org/html/rfc7230#section-5.7.1).
+
+      final _newVia = (targetResponse.headers['via'] != null)
+          ? '${targetResponse.headers['via']}, $_via'
+          : _via;
+
+      resp.headerAdd('via', _newVia);
+      _logProxyResponse.finest('[${req.id}] + via=$_newVia');
+    }
+
+    // Create a stream from the bodyBytes, and use it for the response
+
+    _logProxy.finer('[${req.id}] status=${targetResponse.statusCode}');
+
+    return await resp.addStream(req, () async* {
+      final body = targetResponse.bodyBytes;
+      assert(_size == null || _size == body.length);
+      yield body;
+    }());
   }
 }